backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0